Improved background changing script
The script now stores the previously used backgrounds in a separate
file (already_used) to reduce repetition.

When choosing the next random background, the script reads the file
to see if the background is already there, and if it is - omits it.

If the already_used file contains all the possible images (is full),
a random one is chosen, excluding the last one from already_used
(in order to avoid repetition).